Fix: PyTest failing tests due to name conflict
Upon running the suite of pytests tests I encountered some of the tests
failing. This was due to the `setup` function in
`crhelper_log_helper.py` conflicting with the `pytest` setup module. To
fix this I have renamed the `setup` function to `setupLogger` and
updated all the references to this. I've also updated the classifiers in
`setu.py` to include the latest Python releases